update sdk
ameba-rtos-v1.0 version: 081c4890c6
1. [common][wlan] Change report time of event WIFI EVENT CONNECT
2. [common][wlan] add CONFIG_AMEBAPRO3 which is equivalent to CONFIG_AMEBAGREEN2
3. [common][sscanf] update sscanf minimal
4. [common][wlan] fix sdio bridge release sdk compile issue
5. [amebalite][bt:mesh]mesh model add datatransfer length limit judgement
6. [common][wlan] replace CONFIG_RTLXXXX and RTLXXXX_SUPPORT with CONFIG_AMEBAXXXX
7. [common][soc] add note for nano diag and and scanf
8. [common][bt] add bt inic uart h4/h5
9. [amebasmart][bt]_Modify le iso demo iso data tx flow to avoid invoking sema take action within os timer callback_
10. [amebadplus][sdk] modify cmsis for zephyr cache test
11. [amebasmart][coex] revise sof and eof indication
12. [common][wlan] Add cmd for setting softap ip and ip range in AT+WLSTARTAP
13. [common][example] Fix the compilation warnings related to the examples
14. [common][wlan] Fix comment errors in component/wifi/api/wifi_intf_drv_to_app_basic.h
15. [common][wlan:fw] add rfe_type in h2c_phy_reg_addr
16. [common][atcmd] modify atcmd module
17. [ameba][spi] update example's readme for qc test
18. [common][wlan:fw] replace SYSCFG_RLVersion to EFUSE_GetChipVersion
19. [common][usb] fix imgtool_flashloader build warning
20. [amebadplus][wlan] delete unused ameba_wifi_power_table_usrcfg_1v8_1v25.c
21. [common][usb] switch log api and ic cut version api
22. [common][at_cmd] Fix some issues during the Lwip AT command testing process
23. [smart][wififw]update smart DRFC backup restore table  for 1.0 branch
24. [common][floader] update floader code
25. [amebasmart][hal] fix ota clear error
26. [amebadplus][bt] Refine bt transfer module example
27. [common][sdk] avoid warning for zephyr
28. [common][vfs] replace DiagPrintf by RTK_LOGS
29. [common][vfs] optimize vfs module
30. [common][gdma] update gdma interrupt config function
31. [amebasmart][component] resolve the issue of abnormal disconnection of usbh_wifi bridge during iperf process
32. [ameba][wlan:fw][bug fix] fix high current usage after BT firmware stopped issue

+++ component/soc/amebadplus/fwlib/ram_common/ameba_pll.c
@@ -22,7 +22,7 @@ u32 un_err_normalization;
 bool PLL_Calibration(u32 dst_clk)
 {
 	u32 sys_pll;
-	float des_clk = 0;
+	double des_clk = 0;
 	double target_per;
 	double base_per;
 	double intp_per;
@@ -258,10 +258,10 @@ float PLL_I2S_98P304M_ClkTune(float ppm, u32 action)
 	PLL_TypeDef *PLL = (PLL_TypeDef *)PLL_BASE;
 
 	if (action == PLL_FASTER) {                    /* VCO:688P128M VCO/7=98.304M */
-		F0F_new = F0F_base + (u32)(ppm / step + 0.5);     /* maximum of register is 8191,so ppm<=2719 */
+		F0F_new = F0F_base + (u32)((double)ppm / step + 0.5);     /* maximum of register is 8191,so ppm<=2719 */
 		real_ppm = (double)((double)F0F_new - (double)F0F_base) * step;
 	} else if (action == PLL_SLOWER) {
-		F0F_new = F0F_base - (u32)(ppm / step + 0.5);
+		F0F_new = F0F_base - (u32)((double)ppm / step + 0.5);
 		real_ppm = (double)((double)F0F_new - (double)F0F_base) * step;
 	} else {
 		F0F_new = F0F_base;
@@ -296,10 +296,10 @@ float PLL_I2S_45P1584M_ClkTune(float ppm, u32 action)
 	PLL_TypeDef *PLL = (PLL_TypeDef *)PLL_BASE;
 
 	if (action == PLL_FASTER) {                  /* VCO:677P376M VCO/15=45.1584M */
-		F0F_new = F0F_base + (u32)(ppm / step + 0.5);   /* maximum of register is 8191,so ppm<=3872 */
+		F0F_new = F0F_base + (u32)((double)ppm / step + 0.5);   /* maximum of register is 8191,so ppm<=3872 */
 		real_ppm = (double)((double)F0F_new - (double)F0F_base) * step;
 	} else if (action == PLL_SLOWER) {
-		F0F_new = F0F_base - (u32)(ppm / step + 0.5);
+		F0F_new = F0F_base - (u32)((double)ppm / step + 0.5);
 		real_ppm = (double)((double)F0F_new - (double)F0F_base) * step;
 	} else {
 		F0F_new = F0F_base;
